backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
email_validator==2.1.1  # Added for email validation

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Data Processing
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0  # Fixed version

# Database (optional, for future use)
# sqlalchemy==2.0.27
# alembic==1.13.1
# psycopg2-binary==2.9.9

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.26.0

# Monitoring & Logging
python-json-logger==2.0.7
psutil==5.9.8

# Development
python-dotenv==1.0.1
black==24.2.0
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.9

# Production
gunicorn==21.2.0
prometheus-client==0.20.0